change price to double

+++ src/server/src/main/java/com/media_shop/entity/cart/Cart.java
@@ -15,5 +15,5 @@ public class Cart {
     @Id
     private String id;
     private List<CartItem> listCartItem;
-    private int totalPrice;
+    private double totalPrice;
 }
